+++ CodeChat_Server/CodeChat_Server/__init__.py
@@ -33,3 +33,4 @@ __version__ = "0.0.22"
 #   render_manager.py
 #   renderer.py
 #   __main__.py
+#   constants.py