+++ CodeChat_Server/templates/doxygen/main.c
@@ -1,0 +1,8 @@
+/// \file main.c
+
+#include <stdio.h>
+
+/// The main routine.
+int main(int argc, char** argv) {
+    printf("Hello, world!\n");
+}