+++ docs/CHANGELOG.rst
@@ -14,6 +14,7 @@ Change Log
 -   `Github master <https://github.com/bjones1/CodeChat_system.git>`_:
 
     -   Upgraded to Thrift v. 14.2.
+    -   Added the universal client.
 
 -   0.0.22, 16-Jun-2021: